refactor: extract services test helpers
